Config files should be optional in .component files
Not all VCI components implement a configuration interface so Config
should not be mandatory in the .component files.

Revised plans for how the system boots also makes the need for
defining the config files extraneous. To retain compatibility with
existing components we need to allow the files to be defined, but we
don't need to require this.

Fixes VRVDR-45954