Fix DBUS call via rules file
Part of VRVDR-56271.

When we move to Debian11, we get golang-dbus-5.0.3, and golang-github-coreos-go-systemd 22.x. There's a bug in golang-dbus in that DialHandler() ignores one of the two handlers passed in, causing VCI client method registration to fail. The fix is to change to call Dial() instead using WithHandler / WithSignalHandler() options ... except those don't exist in golang-dbus 4.0.0 which we still need to compile against in Debian10.

Until we drop Debian10 from Vyatta:Unstable, and branch, we need to be able to build (and function) in both Deb10 and Deb11 ISOs. The short-term fix is therefore to make the fix conditional on the package version. As I copied this from the configd equivalent (that got removed a couple of days ago) I have used the coreos-systemd version not golang-dbus, but it essentially comes to the same thing. It's short term anyway - the fix will be put into dbus_transport.go once we drop Deb10 support.

I've verified that the golang-github-vci-dev code pulled into configd / provisiond etc builds is unedited for Debian10, and edited for Debian11. Furthermore, test VCI components build using vci-component-factory appear to work on Debian10 and Debian11.

It would be helpful to get this merged soon so I can properly test in Debian11 as you have to rebuild *everything* that uses VCI to properly test and that is best done by having the base debian11 DANOS:Unstable build having access to this fix.